WAITM-583 Move one off UI into views
